# Fix: rework nondet math operations