## Purpose of Changes and their Description

Re-implement in a deterministic way some math func, here are the details on the chosen algorithms:

- `Exp`: Uses a Taylor polynomial with a lookup table for powers of `10^(1/n)` of size 300 which seems sufficient for 34 precision digits;
- `Ln`: Implements the Newton's method that uses exp 3 times;
- `Log10`: Use `d = ln(x) * 1/ln(10)`;
- `Pow`: Implement `apd`'s algorithm but using our `Ln` and `Exp` impls;

The `Exp10` func has been removed as not used.

The algorithms choices and implementations were made with the idea on improving performances, we are now much faster, see the benchmarks below. In terms of accuracy it is almost identical to previous implementations.

The perf improvements helped the network inferences calculation to drop from 250ms to 100ms.

Benchmarks before:
```
BenchmarkLn/input_1.2-12                    12574	     95334 ns/op
BenchmarkLog10/input_1.2-12              17979	     66241 ns/op
BenchmarkExp/input_1.2-12                  44658	     26917 ns/op
BenchmarkPow/input_1.2_1.2-12          10000	    103102 ns/op
```

Benchmarks after:
```
BenchmarkLn/input_1.2-12                    45258	     26759 ns/op
BenchmarkLog10/input_1.2-12              44979	     27504 ns/op
BenchmarkExp/input_1.2-12                  153902	      7737 ns/op
BenchmarkPow/input_1.2_1.2-12          36795	     32447 ns/op
```

## Are these changes tested and documented?

- [x] If tested, please describe how. If not, why tests are not needed.
- [x] If documented, please describe where. If not, describe why docs are not needed.
- [x] Added to `Unreleased` section of `CHANGELOG.md`?